move VendorTheme our of the model and use theme for every vendor.
