+++ lib/routes/user/add_funds/add_funds_page.dart
@@ -132,7 +132,7 @@ class AddFundsState extends State<AddFundsPage> {
         onTap: () => Navigator.of(context).pushNamed("/fastbitcoins"),
         child: Container(
           decoration: BoxDecoration(
-              color: theme.vendor.fastbitcoins.iconBgColor,
+              color: theme.fastbitcoins.iconBgColor,
               border: Border.all(
                   color: Colors.white, style: BorderStyle.solid, width: 1.0),
               borderRadius: BorderRadius.circular(14.0)),
@@ -143,7 +143,7 @@ class AddFundsState extends State<AddFundsPage> {
                 image: AssetImage("src/icon/vendors/fastbitcoins_logo.png"),
                 height: 24.0,
                 fit: BoxFit.scaleDown,
-                color: theme.vendor.fastbitcoins.iconFgColor,
+                color: theme.fastbitcoins.iconFgColor,
               ),
               Padding(padding: EdgeInsets.only(right: 4.0)),
               Text(
